Refactor quadrature into subfolders
