+++ src/interpolations/quadrilateral.hpp
@@ -2,7 +2,7 @@
 #pragma once
 
 #include "shape_function.hpp"
-#include "quadrature/quadrilateral_quadrature.hpp"
+#include "quadrature/quadrilateral/quadrilateral_quadrature.hpp"
 
 namespace neon
 {
